All Apis with jwt token commit

+++ src/main/java/com/example/demo/utility/JWTUtility.java
@@ -1,5 +1,4 @@
 package com.example.demo.utility;
-
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
@@ -17,7 +16,7 @@ import java.util.function.Function;
 public class JWTUtility implements Serializable {
 
 
-    public static final long JWT_TOKEN_VALIDITY = 5 * 60 * 60;
+    public static final long JWT_TOKEN_VALIDITY = (long) 5 * 60 * 60;
     private static final long serialVersionUID = 234234523523L;
     @Value("${jwt.secret}")
     private String secretKey;
